+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/deploy
 
+## 0.18.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.18.3
+
 ## 0.18.2
 
 ### Patch Changes