+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/doc
 
+## 0.18.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.18.3
+  - @rocketh/node@0.18.6
+
 ## 0.18.5
 
 ### Patch Changes