+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocketh/export
 
+## 0.18.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.18.3
+  - rocketh@0.18.5
+  - @rocketh/node@0.18.6
+
 ## 0.18.5
 
 ### Patch Changes