+++ packages/rocketh-read-execute/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/deploy
 
+## 0.18.4
+
+### Patch Changes
+
+- add confirmationsRequired option
+- Updated dependencies
+  - @rocketh/core@0.18.3
+
 ## 0.18.3
 
 ### Patch Changes